solver_based_model.py refactored to core/solver_based_model.py, core/reaction.py, and core/solution.py
